Added sort assigned customers by index to MutableSolution.update()
When we have a target centre and 2 customers, the most central customer
is arbitrary (either has equal cost) and is determined by the order of
customers in the assigned list. This can create problems, as solution
updates can give different costs dependent on the ordering in the
customer list. This can cause an infinite loop in the local search where
we search until no improvement. We therefore sort customers by index so
the update becomes deterministic.